requirements.txt
pandas
win32com.client
sys
datetime
os
xlwings
numpy
xlrd
xlwt
openpyxl